Fixes: NB#154090 Localization of the applet loaded too late
Details:

Before only the translation of the currently visible applet was loadedr,
but in this case the translations of the briefs should be part of controlpanel
instead of the applet. Which is not working, at least for applets made by
externals.

So I restored the old behaviour: loading all the translations at startup,
with the exception that it takes into account the catalog from the desktop file.