Fixes: NB#165815 - Duicontrolpanel calls title ...
function for an applet wrongly

Details:

Setting the locale generated retranslateUi() calls which made the title
of the appletpage refreshed to the value which DcpAppletIf::title() call
provided (the title of the just openned page). Fixed by only setting the
locale when new catalogs are added.

To avoid the need applet developers have to remember the id of the last
accessed page etc we will bring the DcpAppletIf::title() call into the
widget itself: DcpWidget::title(). This also means deprecating the
DcpAppletIf::title() call to avoid confusion.