Fixes: NB#169944 - Recently Used list in DCP is sorted improperly
Details:

The recently used item list should put the last clicked applet to the
top, but also has to respect the usagecount of the applets. This patch
does that the last clicked item gets handled separately, always goes to
the top, and the other 3 items are the most used ones.