Fixes: NB#172559 - Time and Language UI display some hard coded button (Display language and Keyboard languages)
Details:
MContentItem has an alreadyfixed bug (only not integrated) that
events are not coming for the widgets, including the event
which calls retranslateUi, which previously translated the titles.

I changed the behaviour so, that catalogs are already loaded for the
applet by the time when the widget gets created, which is probably wiser
anyway.