Fixes: NB#173756 - The order of the items in recently used list is not same after the device reboot
Details:
The gconf key for the last clicked item was saved only when the application was
closed. Therefor if the user rebooted without closing the controlpanel, the
last clicked item was lost, while the most useds are saved on every
click, so they remained.

Modified the behaviour so that now at every applet openings the value
gets saved, I guess it is not a big overhead.