Fixes: NB#230541
Details:
The page request signals of the applet did not get connected in case the
applet was started through dbus directly. The cause turned out to be
that the PageFactory was created _after_ the applet, so it did not get
the appletLoaded signal from the db.

Fixed by iterating through already loaded applets on PageFactory creation
(it should not be dependant on the creation sequence.)