Brightness applet has different title if invoked through DBus
Details:
Fixes: NB#255038 - Settings; while searching for the string Brigtnees
will be shown as Display in Settings view

The cause was that the applet title is not propagated through the
localsocket if the brief is correctly running in the briefsupplier
process. Thus, in this normal case the applet's title is coming from the
.desktop file.

But, when the applet was started by DBus, controlpanel's main
process forced loading the applet (for behaving like an appletlauncher),
so its brief was used directly, resulting in using the deprecated
DcpAppletIf::title() for translating the string.