Fixes: NB#258212 - On reopening the settings, ...
the settings app is getting minimized

Details:
The cause is a timing issue, if the window closing start of the applet
launcher instance happens after the main controlpanel window raising,
then the letter gets overridden because of the animation (the applet launcher
window will be the raised one).

Fixed it by openning a new window even if the same page gets requested
again by servicefw, because it also fixes Bug NB#258395 (View
transition missing from Settings app when opened from Launcher and
already open in Recents).