fixed 267658 - Opening Bluetooth from Status menu is very slow
and shows bad transition

Details:
When a DBus request comes to pop up an appletpage control panel has to
drop each of its pages (which means dropping some processes). It also
has to be taken into account that each applet should have a different
process.

So how it achieved this earlier was that on DBus request if already
displaying something
- first it closed down every applet launcher process
- second unregistered itself from DBus
- started itself again to achieve the applet process separation

I modified this behaviour, because 99% sure an empty appletloader
process is waiting and gets closed in the above way. So instead, that
preloaded appletlauncher process takes the role to display the newly
requested page and only the others close down including the main.