Fixes: NB#287787 - does not open, if tapped immediately
Details:
Sounds and vibration does not open, if tapped immediately after
opening Settings view.

Control panel prestarts an applet launcher process with a hidden
window after startup finished to speed up applet opening.
The reason for this bug was that the Sounds applet got the
appletlauncher DBus prestart request, which resulted in hiding
the appletlauncher's window.