Prevents handling more applet requests in the appletlauncher
Details:
Fix for Bug NB#289899.
If an applet request came twice to the appletlauncher when the desktop file
were already loaded, the second trial to load the desktop file again
returned false (== failure), so the process has closed assuming the
applet could not be found.

(Happens only during stress testing.)