Fixes: NB#292482 - "Internet connection" UI flashes back ...
to "Settings" UI when it has been open again from "settings".

Details:

Main process gets the DBus request that the user wants the mainpage.
So sends the quit signals to all appletlauncher process (even the
preloaded one).
Then the preloaded process gets the applet request, shows the page
for a blink, but is also shutting down because of the quit signal.

Solved it by not wasting the preloaded process, because it is
unnecessary.